+++ src/pages/SignUpPage/SignUpPage.tsx
@@ -1,7 +1,5 @@
-import { useForm } from 'react-hook-form';
 import { CheckIcon } from '@radix-ui/react-icons';
-import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
+import SignUpCompletePopup from '../../components/popUp/SignUpCompletePopup.tsx';
 import {
   Container,
   HeaderContainer,
@@ -23,16 +21,8 @@ import {
   HintMessage,
   DuplicateCheckButton,
 } from './SignUpPage.style.ts';
-import SignUpCompletePopup from '../../components/popUp/SignUpCompletePopup.tsx';
-import { PATH } from '@/constants/path';
-
-interface FormData {
-  name: string;
-  email: string;
-  password: string;
-  confirmPassword: string;
-  agreement: boolean;
-}
+import { useSignUpForm } from '../../../src/hooks/signUp/useSignUpForm.ts';
+import { useNavigate } from 'react-router-dom';
 
 const SignUpPage = () => {
   const navigate = useNavigate();
@@ -40,36 +30,18 @@ const SignUpPage = () => {
   const {
     register,
     handleSubmit,
-    watch,
-    formState: { errors, isSubmitting, isValid },
-  } = useForm<FormData>({
-    mode: 'onChange',
-  });
-
-  const [isEmailChecked, setIsEmailChecked] = useState(false);
-  const [isAgreeChecked, setIsAgreeChecked] = useState(false);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const password = watch('password');
-
-  const onSubmit = (data: FormData) => {
-    console.log('회원가입 데이터:', data);
-    setIsModalOpen(true);
-  };
-
-  const closePopup = () => {
-    setIsModalOpen(false);
-    navigate(PATH.EDIT_PROFILE('1'));
-  };
-
-  const checkEmailDuplicate = () => {
-    console.log('이메일 중복 확인');
-    setIsEmailChecked(true);
-  };
-
-  const checkAgreeButton = () => {
-    console.log('동의 버튼 체크 확인');
-    setIsAgreeChecked(true);
-  };
+    errors,
+    isSubmitting,
+    isValid,
+    isEmailChecked,
+    isAgreeChecked,
+    isModalOpen,
+    onSubmit,
+    closePopup,
+    checkEmailDuplicate,
+    checkAgreeButton,
+    password,
+  } = useSignUpForm();
 
   return (
     <Container>
